LearnAPI/project/requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.4
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
Flask-HTTPAuth==4.2.0
Flask-JWT==0.3.2
tinydb==4.3.0
